Fix the search in ResultsDB
Fixes few hiccups in the search implementation.
First of all, the source for typeahead.js is changed to CDN that
serves the data over https, instead of http (which was causing problems
on prod).
Also, the testcase_tokenizer endpoint works if there are multiple pages
of testcases.
Search form is populated with the data from the URL, if present.

Differential Revision: https://phab.qa.fedoraproject.org/D1106